ITH_PipelineS_UpdateSeuratToV3
Updated the ITH Pipeline S to account for syntax changes to Seurat CRAN repository release v3.0. This allows users to have an expanded toolset of analysis of single-cell RNA sequencing datasets in addition to the generalized diversity calculations included in this pipeline. Here is a command list cheat-sheet for differences in Seurat commands and structure from v2.X and the updated v3.X: https://satijalab.org/seurat/essential_commands.html